requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
asyncpg==0.30.0
PyQt5==5.15.10
PyYAML==6.0.1
qasync==0.24.0
pydantic==2.5.0 
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0 